requirements.txt
# ═══════════════════════════════════════════════════════════════════
# TRADING BOT PROFISSIONAL - DEPENDÊNCIAS
# ═══════════════════════════════════════════════════════════════════

# Core
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Exchange
ccxt>=4.2.0
requests>=2.31.0

# MetaTrader5 (Windows)
MetaTrader5>=5.0.45

# API (FastAPI)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Database
# sqlite3  # Built-in
